requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.8.0
bleak==0.13.0
bleak-winrt==1.2.0
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-autoinstall==0.0.7
click==8.1.3
colorama==0.4.6
cycler==0.10.0
Django==3.2.12
dynamixel-sdk==3.7.31
exceptiongroup==1.0.1
future==0.18.2
h11==0.14.0
idna==3.4
importlib-metadata==5.2.0
imutils==0.5.4
iso8601==0.1.12
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.1.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.1.1
mpmath==1.2.1
nest-asyncio==1.5.4
numpy==1.16.2
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
outcome==1.2.0
Pillow==6.1.0
pip==22.3.1
pycparser==2.21
pymodi==1.2.0
pyparsing==2.4.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2021.3
PyYAML==5.1.2
requests==2.28.1
selenium==4.6.0
setuptools==65.6.3
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==1.9.2
sqlparse==0.4.2
sympy==1.10.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.12
virtualenv==16.7.7
websocket-client==1.2.3
Werkzeug==2.2.2
wheel==0.38.4
wsproto==1.2.0
zipp==3.11.0
